// sdk/src/types.ts
export interface TokenMeta {
  id: string;
  name: string;
  symbol: string;
  decimals: number;
  total_supply?: string;
  creator?: string;
  created_height?: number;
  created_time?: number;
  create_txid?: string;
  transfer_count?: number;
  verified: boolean;
  logo_url?: string;
}

export interface TokenTransfer {
  txid: string;
  block_height?: number;
  block_time?: number;
  from_address?: string;
  to_address?: string;
  amount: string;
  confirmations?: number;
}